Bitte Herr Schoch

+++ server/Packet.h
@@ -2,7 +2,7 @@
 #pragma once
 typedef struct Packet {
 	char txtCol[256];
-	int seqNr;
+	unsigned short seqNr;
 	unsigned short checkSum;
